Create a project laid out for producing training materials

Create a project laid out for producing training materials

createTrainingProject(name, dirs = c("data", "handouts", "slides"),
  handoutEngine = "rmarkdown", slideEngine = "rmarkdown", ...)

Arguments

name
Project

dirs
Directories to create

handoutEngine
Package for building handouts

slideEngine
Package for building slides

...
Pass to createBasicProject()

Examples

# NOT RUN {
proj<-"trainingproj"
createTrainingProject(proj)
list.files(proj)
unlink(proj)
# }